test: relocate EVPN converger helpers
